// src/app/interfaces/hospedajes.interface.ts
export interface Hospedaje {
        created_at: string;
        deleted_at: string;
        descripcion: string;
        dias?: number;
        id: number;
        image:  string;
        precio: number;
        tipo:  string;
        updated_at: string;
        image0: string;
        image1: string;
        image2: string;
}